Edit profile pic modal seperated
